Basic Teleportation
The new scene should be named "DevelopmentEnvironment". To test teleportation, open the scene then press play. I configured the VRTK simulator using a bezier pointer renderer...I tried using straight pointers but I could not figure out how to aim the simulated controllers with these. It may well be that you have to use a real headset to test the straight laser pointers. The controls for the simulator in the scene are: WASD for movement, mouse for looking, Q to fire laser/teleport, and tab to switch hands. Green pointer = valid location, red = invalid. If there are any issues let me know asap.