Pointer Selection Initial Implementation
I have a written a script named "ObjectPointerInteraction.cs" that triggers a Debug.Log() statement when the straight laser pointer touches a VRTK interactable object. The log prints the name of the object that is touched by the laser, and only the name of that specific object. Only objects that have this script AND the VRTK Interactable Objects script attached as components can trigger this print statement.