Initial Implementation of Road.cs
Road Class contains a doubly linked list (type LinkedList in C#) that contains the lanes currently in the road. It has an insertLaneAtEnd function that places a new lane on the outside of the lane that was previously placed. This function is called twice in the Start function to create the two default starting lanes. Have not implemented remove, any pointer interaction with lanes, and have unsuccessfully tried to unit test the Road.cs file.