Road Class Stubs
Stubbed out BasicLane and all lane subclasses, BasicStripe, and Road. BasicLane is basically good to use for testing other parts of the program. Road.cs needs more work. When programming this class, we need to make sure to make use of the BasicLane functions as much as possible in order to simplify it, because otherwise the class will just get too complicated.