Stripe placement partially implemented
Added a lot of new functionality to Road.cs, BasicLane.cs, and Stripe.cs. The one that needed the most work was Road.cs; I had to add a rather large helper function for insertLane that dealt with stripe placement (there are four substantial cases when it comes to stripe placement in a new lane). This function is called setStripes(), and is a private void. I also added a function to Stripe.cs that properly positions a single stripe, and it is called setStripePosition(). Finally, I added a setLaneOrientation(string orientation) to Stripe.cs that sets the left or right lane of the stripe (depending on the specified orientation) and I added a function to BasicLane.cs called setStripeOrientation() which sets the left or right stripe of the lane. All in all, very substantial commit, but the stripes still definitely need more work.