Stripe Functionality Implemented
I have finished implementing the helper function that deals with stripes for insertLane(). I also added a similar helper function for removeLane() called resetStripes() and made a tiny tweak to adjustRoadAroundLane() and setLaneWidth() that changes the way the road adjusts the strripes' positions when lane width is changed. The reason for this is that there was a bug in the old system with our new stripe setup. I tested out all of the cases I could think of and it all seems to work, but please feel free to test it yourselves. If you have any questions please ask!