Sidewalks implemented
I have added a function in Road.cs for changing a lane's type. As it turns out, the way I originally envisioned changing a lane's type (from inside BasicLane.cs) simply won't work because we don't have authorization to change the gameObject variable. However, I was able to work around this using insertLane(), setLaneWidth(), and removeLane() in the road.cs setLaneType(). Pressing the drop down menu on the manipulation menu and selecting one of the options should now insert a sidewalk.