All lane type skeletons implemented
All of the lane types have been implemented with very basic appearances. They have full functionality with regards to stripe placement and size. The only work that really remains is creating the textures for each of the lanes and creating the icons that go on each lane (e.g. bike symbol for bike lane).