Set Stripe Type function implemented
I have successfully implemented a method for changing the stripe type model of an instantiated stripe. I have changed the structure of stripe objects a bit (the script is now attached to a container which holds the stripe model). There is currently an insertion bug with different stripe types in which an altered stripe will not stay on the lane it is attached to (if you insert to the right of the altered stripe). This should be an easy fix.