Removed handleVehicleLaneStripes()
I removed this helper function for setLaneType() because it adds a lot of annoyingly complex cases. I did not totally delete it in case we want to add it back in later, but for now it is commented out.
What we have so far on this branch is all of the lane types, a function that changes lane types, proper stripe shifting, insertion, and deletion in every case, and finally a function that can change a stripe's type (located within stripe.cs).
While this code runs, I may have to go back later and clean it up. I would definitely appreciate some code reviews of road.cs; the file is nearing a thousand lines now and seems a bit excessive (though there are a lot of cases). If there is any way I could condense this code, I am open to suggestions.