Implemented Saving for the Road
I have written a few new scripts that handle saving, and I have started work on loading as well. So far, I have successfully saved the data in our road (although I have not loaded it back in yet, I'm still working on that). In a nutshell, I am storing all of the important class variables in road, BasicLane, and Stripe and putting it in a binary file.