Implemented Saving and Loading for Buildings and Fog
The application's save system can now save the current buildings and fog distance to a binary file in addition to all of the road's lanes. All of these variables can then be loaded back in exactly as they were saved. Saving and loading is almost done being implemented; we only need to implement it for props now.

Note: there is still a bug in the Start() function for Road.cs. This bug is related to updateBuildings(), although I am not sure which part of the program is causing the problem here. Basically, there is a gap between the buildings and the road that should not be there, resulting in a view of the infinite drop below. Fortunately this does not hamper development of saving and loading, but it should be fixed asap.