Code for "no stripe" option.
I altered the setStripeType() function in Stripe.cs so that if a newType is passed in that does not match any of the stripe prefabs, it just destroys the existing stripe and doesn't make a new one. I also wrote a fix for that stripe bug in loading by making sure the saved stripe data is set to null during save.